requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Django==4.0.3
django-cors-headers==3.11.0
django-redis-cache==3.0.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
mpmath==1.2.1
numpy==1.22.3
patool==1.12
Pillow==9.1.0
pycparser==2.21
pynput==1.7.6
python-decouple==3.6
pytz==2022.1
pywin32==303
rarfile==4.0
redis==3.5.3
requests==2.27.1
scipy==1.8.0
six==1.16.0
sounddevice==0.4.4
sqlparse==0.4.2
sympy==1.10.1
tzdata==2022.1
unrar==0.4
urllib3==1.26.9
